From bug #1389723 Roadmap for 0.92: (this bug) ================= TODO Non-pixel units are allowed but everything in a document should use the same units including guides and grids. TODO (Guides and grids need code to do this.) DONE An extension that saves a document in pixels should be added to allow backwards compatability with 0.91.
It should be possible to avoid overwriting Inkscape version until any necessary conversion is done. If a conversion appears to be needed, a pop-up can ask the author if it should be done.

Roadmap for 0.91: (bug #1387864)
=================
TODO: Everything should use only pixels as is (mostly) done in 0.48 (including templates).
TODO: The ability to switch document units should be disabled (it's broken in several ways).
TODO: Font size should be stored in pixels.
DONE: Some templates in 0.48 templates do use 'mm' so there does need to be a script to convert files based on those templates to 96dpi.

(bzr r13693)