Extensions: try to calculate the SVG unit
We cannot use inkscape:document-unit for the SVG unit, because it specifies the UI display unit in Inkscape and is unrelated to the units used for SVG values.
Instead, the real units of the SVG values is defined by document height/width and the viewbox. The calculation assumes isotropic units (give me a break!) and only looks at the width. It tries to match that with a list of known units.
If no match found: assume 'px' units.
Finds 'px' units for legacy Inkscape documents without viewbox.

(bzr r13750)