Add fonts tab to svgcalender extension
This commit adds a crude Fonts tab and option to choose fonts for
strings in the calendar.

Currently, calendar render does not include chosing custom fonts
for strings for Year, Month names, Day of the week names, and day
names. This change gives the user an option to manually input the
names of fonts they want to use for these strings.

The default font for the text is the generic Arial font family.
Users can type the name of a different font. Ideally, this string
field will be replaced soon with a drop-down font selector.

Hopefully, this change will make the calendar render extension
more useful and customizable for everyone.