Fix Unicode output for Text -> Extract extension
This prevents errors occurring if text contains non-ASCII characters.
This also fixes Unicode support for AbortExtension.

Note that Inkscape must explicitly request UTF-8 output with
PYTHONIOENCODING, otherwise the text will just be backslash escaped.

Also fix tests to catch this in the future. It appears this was caused
by a regression in commit 7fea47c59300 that made all strings Unicode.

Life would be easier if we just supported Python 3 :(

Fixes: https://gitlab.com/inkscape/extensions/issues/76