Fix path transform with relative command after "z"
After encountering a "z" command, Path.transform() continued at the
closed path's untransformed start coordinates both in untransformed and
transformed space, messing up any relative command that followed.

This change keeps track of the transformed starting point of the current
sub-path (in addition to the untransformed) so that "z" can move the
current position to the correct place in transformed space, too.