only return elements we know how to treat and ignore other unknown elements
Prior to this commit, if get() was called on a selection with embedded comments in the SVG it failed.

To resolve this issue, this commit filters the returned elements for BaseElement and str and returns
only those as it could very well be possible that the selection contains other unknown elements.

Fixes: #369

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>