Proper CSS selectors implementation
This commit replaces the previous implementation - which required
O(n) evaluations, where n is the number of elements in the document,
simply to check whether a given element has a class - with a proper
O(1) implementation. It is a fork on cssselect2 which doesn't use eval
and doesn't require duplication of our tree structure.

This should be mostly compliant with CSS Selectors Level 4.

The ConditionalStyle.to_xpath method has been deprecated and replaced
with ConditionalStyle.all_matches.