Fix Windows non-ANSI dxf export crash
Before the change, there was a Windows-specific crash reported on exporting an SVG containing text with superscript 5 (⁵) as well as certain Chinese characters to dxf.
These can be generalised as non-ANSI characters.

This change fixes the crash.

Issues solved:
inkscape/extensions#593
inkscape/extensions#586

This is a forward port of MR:
inkscape/extensions!678